WIP: upgrade speed param from int to float

+++ src/core/image_utils.rs
@@ -145,12 +145,13 @@ pub trait SpeedOptimizer {
     fn reference_cache(&self) -> Self::ReferenceCache;
 
     /// Modifies the parameters of the image in-place.
-    /// An optimization level of zero corresponds to the "default paramers", with positive
-    /// integers corresponding to progressively faster render times (and thus lower quality).
-    ///
+    /// An optimization level of zero corresponds to the "default paramers" and one corresponds 
+    /// to "as fast as possible, with dramatic loss of image quality".  It is up to each fractal
+    /// implementation to anchor the upper bound to a meaningful value.
+    /// 
     /// Note: parameters modified by this call should strictly reduce the render time, and
     /// should not change the size of the image or underlying data structures.
-    fn set_speed_optimization_level(&mut self, level: u32, cache: &Self::ReferenceCache);
+    fn set_speed_optimization_level(&mut self, level: f64, cache: &Self::ReferenceCache);
 }
 
 /// Scales down an integer parameter based on a scale factor.
@@ -195,9 +196,10 @@ impl SpeedOptimizer for RenderOptions {
         *self
     }
 
-    fn set_speed_optimization_level(&mut self, level: u32, cache: &Self::ReferenceCache) {
+    fn set_speed_optimization_level(&mut self, level: f64, cache: &Self::ReferenceCache) {
         self.downsample_stride = cache.downsample_stride + (level as usize);
 
+        //MPK:  TODO -- make a better scaling law here, with a float to int cast.
         self.subpixel_antialiasing = if cache.subpixel_antialiasing > level {
             cache.subpixel_antialiasing - level
         } else {
